@@ -19,11 +19,11 @@ Gem::Specification.new do |s|
 
   s.add_dependency 'rails'
 
-  s.add_development_dependency 'rails', '~> 5.0.0', '>= 5.0.0.1'
+  s.add_development_dependency 'rails', '~> 5.2.0'
   s.add_development_dependency 'sqlite3'
   s.add_development_dependency 'pry'
   s.add_development_dependency 'rspec-rails'
-  s.add_development_dependency 'factory_girl_rails'
+  s.add_development_dependency 'factory_bot_rails'
   s.add_development_dependency 'database_cleaner'
   s.add_development_dependency 'faker'
 end
