@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-FactoryGirl.define do
+FactoryBot.define do
   factory :boolean, class: FriendlyRoutes::Params::BooleanParams do
     statements { Faker::Lorem.words(2) }
 
