@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-FactoryGirl.define do
+FactoryBot.define do
   factory :collection, class: FriendlyRoutes::Params::CollectionParams do
     transient do
       items 0
